Added frost attack
Lectern now grants the player a frost attack. Changed the gate from being a "door" to an obstacle that the user has to get past so they can exit the level. Added some new art. The level is now ready to be put together with only a few more additions to go.